Fix test runner bug which won't update bundle
* Consider this scenario: _tmp directory holds a particular version of
 cucumber/cucumber and some commits were made to the remote of
 cucumber-ruby-core. Now, after running git pull in the
 cucumber-ruby-core repo to fetch the latest changes, running the
 integration tests would fail since the Gemfile.lock inside
 cucumber/cucumber stores a particular commit SHA from the previous run
 of integration spec.

* To get around this issue, `bundle update cucumber-ruby-core` inside
 cucumber/cucumber should be run after pulling the changes in
 cucumber-ruby-core. But this will fail if `bundle install` is not run
 in the first place. So, the script would check for the presence of
 Gemfile.lock and run `bundle update` or `bundle install` accordingly.

+++ Rakefile
@@ -12,14 +12,22 @@ RSpec::Core::RakeTask.new(:spec) do |t|
 end
 
 namespace :integration do
-  cucumber_ruby_core_folder = Dir.pwd
+  desc "Equivalent to running rake inside cucumber/cucumber"
+  task :spec => [:bundle_update]  do
+    sh "bundle exec rake"
+  end
 
-  task :spec => [:pull_cucumber] do
-    sh "rake build"
+  task :bundle_update => [:pull_cucumber] do
+    cucumber_ruby_core_folder = Dir.pwd
     cd "_tmp"
     sh "bundle config local.cucumber-core #{cucumber_ruby_core_folder}"
-    sh "bundle install --path=vendor"
-    sh "bundle exec rake"
+    if File.exists?("Gemfile.lock")
+      puts "A Gemfile.lock exists. Update cucumber-ruby-core commit"
+      sh "bundle update cucumber-core"
+    else
+      puts "No Gemfile.lock found. Installing all gems"
+      sh "bundle install --path=vendor"
+    end
   end
 
   task :pull_cucumber => "_tmp" do
